requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
alembic
psycopg2-binary
python-docx
pdfplumber
pytesseract
pillow
groq
jinja2
aiofiles
python-dotenv
aiofiles
alembic
fastapi
groq
jinja2
pdfplumber
pillow
psycopg2-binary
pytesseract
python-docx
python-dotenv
python-multipart
sqlalchemy
uvicorn
aiofiles
alembic
fastapi
groq
jinja2
pdfplumber
pillow
psycopg2-binary
pytesseract
python-docx
python-dotenv
python-multipart
sqlalchemy
uvicorn
aiofiles
alembic
fastapi
groq
jinja2
pdfplumber
pillow
psycopg2-binary
pytesseract
python-docx
python-dotenv
python-multipart
sqlalchemy
uvicorn